Add way to add custom arguments to the Xcode plugin
Looks for the first line in the main configuration file containing the label `"xcode-plugin-arguments:"`. If found, it treats the rest of the line after the label as command line arguments and adds them to the tool invocation.

This can be used, for example, to selectively turn on `—strict` linting.

Note that while adding `—fix` will work, Xcode build plugins don’t have write access to their input files and so the fixes will end up in the plugins output working directory and not in the original files.

For example, if the main config file looks as follows, this will add `--strict` “and `—no-cache` to the tool invocation run by the plugin.

```
# xcode-plugin-arguments: --strict --no-cache

included:
   - Sources
   - Tests
```